Sign Up logic is better in Flux file
