+++ src/front/js/pages/signUp.js
@@ -1,9 +1,11 @@
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 import { useHistory } from "react-router-dom";
+import { Context } from "../store/appContext";
 // Bakend URL
-import { API_BASE_URL } from "../constants";
+//import { API_BASE_URL } from "../constants";
 
 export const SignUp = () => {
+	const { actions } = useContext(Context);
 	// to go "Edit Profile" after the sign up
 	let history = useHistory();
 	const [formValue, setFormValue] = useState({
@@ -21,31 +23,16 @@ export const SignUp = () => {
 
 	const handlerSubmit = e => {
 		e.preventDefault();
-
-		// To create a new user directly in DB
-		// for this reason the new user does not to be save in the Flux
-		const raw = JSON.stringify(formValue);
-
-		const requestOptions = {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: raw,
-			redirect: "follow"
-		};
-
-		fetch(`${API_BASE_URL}/api/sign_up`, requestOptions)
-			.then(response => response.json())
-			.then(result => {
-				console.log("New user was created: ", result);
-				// to go "Edit Profile" after the sign up
-				history.push("/edit_profile");
-				//setFormValue({ [e.target.name]: "" });
-			})
-			.catch(error => console.log("error", error));
+		actions.signUp(formValue);
+		// to go "Edit Profile" after the sign up
+		history.push("/edit_profile");
 	};
 
 	return (
 		<div className="container row text-center mt-5  d-flex justify-content-center">
+			<div>1</div>
+			<div>2</div>
+			<div>3</div>
 			<form onSubmit={handlerSubmit} className="col-6 p-5 mt-3 myBox text-white">
 				<div className="form-grup row mt-2">
 					<input
@@ -74,7 +61,7 @@ export const SignUp = () => {
 						className="col-6"
 						type="password"
 						name="password"
-						placeholder="Enter yor password"
+						placeholder="Password"
 						onChange={inputHandelChange}
 						required
 					/>
